backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0
pandas>=2.0
jinja2>=3.1
python-multipart
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0
bcrypt>=4.0.0
PyJWT>=2.8.0
xhtml2pdf>=0.2.13
